warpimage - updating to accept a numpy array
initial commit.
Modify warpimage to accept a numpy array as input image. This allows the user to add warpimage to a chain of image processing instead of simply opening from an existing image file.
Example: We have a global topo map to which we want to apply a custom colormap. Instead of reading, modifying, writing, re-reading, and re-projecting... we can read and modify using PIL calls, and re-project with warpimage.